Minor changes to scripts for production deployment

+++ workoutapp/prodstop.sh
@@ -1,3 +1,7 @@
-pid=$(cat /var/run/gunicorn/prod.pid)
-echo $pid
-kill $pid
+#!/bin/bash
+
+if [ -f "/var/run/gunicorn/prod.pid" ]; then
+    pid=$(cat /var/run/gunicorn/prod.pid)
+    echo $pid
+    kill $pid
+fi
